+++ src/onegov.election_day/onegov/election_day/tests/test_vote_views.py
@@ -40,6 +40,7 @@ def test_view_vote_summary(election_day_app):
 
         assert client.get('/vote/vote/summary').json == {
             'answer': 'rejected',
+            'completed': True,
             'date': '2015-01-01',
             'domain': 'federation',
             'last_modified': '2014-01-01T12:00:00+00:00',